fix(worldgen-v4): 清理 CodeRabbit nitpick（去引号注解/冗余 int(round)/f-string/无用 import）
- fields.py from_slots 返回注解去引号（文件头已有 from __future__ import annotations）
- stitcher / regenerate_v3_baseline / test_v3_behavior_baseline 去掉 int(round(...)) 的冗余 int()
  （Python3 round 无第二参返回 int）
- test_v3_behavior_baseline neg_pressure 断言 .format() 改 f-string
- test_spans 移除未使用的 ColumnSpans 本地 import

+++ worldgen/scripts/terrain_gen/stitcher.py
@@ -256,7 +256,7 @@ def blend_spans(
     structural = overlay_spans if weight >= 0.5 else base_spans
     base_ceiling = base_spans[0][1]
     ov_ceiling = overlay_spans[0][1]
-    blended_ceiling = int(round(base_ceiling + (ov_ceiling - base_ceiling) * weight))
+    blended_ceiling = round(base_ceiling + (ov_ceiling - base_ceiling) * weight)
 
     struct_floor = structural[0][0]
     # Guard: the lerped surface must stay above the structural span's floor.
